service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
prometheus-client==0.22.1
pydantic==2.11.7
scikit-learn==1.7.1
pytest==8.4.1
httpx==0.28.1
python-dotenv==1.1.1
